+++ Controllers/StockController.cs
@@ -41,7 +41,7 @@ namespace api.Controllers
         }
 
         // GET: api/stock/{id}
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
             try
@@ -64,6 +64,7 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if (!ModelState.IsValid) return BadRequest();
             var stockModel = stockDto.ToStockFromCreateDTO();
             await _stockRepo.CreateAsync(stockModel);
 
@@ -72,10 +73,10 @@ namespace api.Controllers
 
         // PUT api/stock/{id}
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
-
+            if (!ModelState.IsValid) return BadRequest();
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
 
             if (stockModel == null)
@@ -87,9 +88,10 @@ namespace api.Controllers
 
         // DELETE api/stock/{id}
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest();
             var stockModel = await _stockRepo.DeleteAsync(id);
 
             if (stockModel == null)
